add first getlocation test
add utilities for GetLocationApiTests
dramatically reduce code duplication in response parsing/deserializing
make lat/long vals in Locations enum/pojo be stored as doubles

+++ WebAppAutoTesting/src/main/java/api/responses/Location.java
@@ -18,6 +18,6 @@ protected String cityId;
 protected String stateId;
 protected String countryId;
 protected int woeid;
-protected int latitude;
-protected int longitude;
+protected double latitude;
+protected double longitude;
 }
